requirements.txt
Pillow==6.1.0
python-telegram-bot==12.0.0b1
requests
